requirements.txt
Django==1.10.4
psycopg2==2.6.2
djangorestframework==3.5.3
django-filter==1.0.1